Move interactor to the Domains namespace
