--- app/interactions/domain_update_confirm_interaction/process_action.rb
@@ -1,15 +1,0 @@
-module DomainUpdateConfirmInteraction
-  class ProcessAction < Base
-    def execute
-      ::PaperTrail.request.whodunnit = "interaction - #{self.class.name} - #{action} by"\
-        " #{initiator}"
-
-      case action
-      when RegistrantVerification::CONFIRMED
-        compose(ProcessUpdateConfirmed, inputs)
-      when RegistrantVerification::REJECTED
-        compose(ProcessUpdateRejected, inputs)
-      end
-    end
-  end
-end
