--- app/interactions/domain_update_confirm_interaction/process_update_rejected.rb
@@ -1,16 +1,0 @@
-module DomainUpdateConfirmInteraction
-  class ProcessUpdateRejected < Base
-    def execute
-      ActiveRecord::Base.transaction do
-        RegistrantChangeMailer.rejected(domain: domain,
-                                        registrar: domain.registrar,
-                                        registrant: domain.registrant).deliver_now
-
-        notify_registrar(:poll_pending_update_rejected_by_registrant)
-
-        preclean_pendings
-        clean_pendings!
-      end
-    end
-  end
-end
