+++ app/jobs/domain_update_confirm_job.rb
@@ -3,8 +3,8 @@ class DomainUpdateConfirmJob < ApplicationJob
 
   def perform(domain_id, action, initiator = nil)
     domain = Epp::Domain.find(domain_id)
-    DomainUpdateConfirmInteraction::ProcessAction.run(domain: domain,
-                                                      action: action,
-                                                      initiator: initiator)
+    Domains::UpdateConfirm::ProcessAction.run(domain: domain,
+                                              action: action,
+                                              initiator: initiator)
   end
 end
